requirements.txt
# ============================================================================
# LangChain 1.0 学习仓库 - Python 依赖文件
# ============================================================================
# 使用说明：
#   pip install -r requirements.txt
# 或使用 uv（更快）：
#   uv pip install -r requirements.txt
# ============================================================================

# ----------------------------------------------------------------------------
# LangChain 核心包（1.0 版本）
# ----------------------------------------------------------------------------

# LangChain 核心库
langchain>=1.0.0,<2.0.0
langchain-core>=1.0.0,<2.0.0
langgraph-cli[inmem]

# 群组agent
langgraph-swarm

# 主管agent
langgraph-supervisor

# LangChain Classic - 包含部分移至经典包的组件（如 EnsembleRetriever）
langchain-classic>=1.0.0

# LangGraph - Agent 运行时
langgraph>=1.0.0,<2.0.0

# LangSmith - 可观测性平台（可选但推荐）
langsmith>=0.2.0

langchain-mcp-adapters
fastmcp=2.13.1

# ----------------------------------------------------------------------------
# 模型集成包（按需选择）
# ----------------------------------------------------------------------------

# ⭐⭐⭐ Groq 集成（推荐！免费且超快）
langchain-groq>=0.2.0

# OpenAI 集成
langchain-openai>=0.2.0

# Anthropic 集成（Claude）
langchain-anthropic>=0.3.0

# Google Gemini 集成（免费额度慷慨）
langchain-google-genai>=2.0.0

# Cohere 集成（可选）
# langchain-cohere>=0.3.0

# HuggingFace 集成（用于 Embeddings - Module 13, 14）
langchain-huggingface>=0.1.0

# HuggingFace Transformers 和 Embeddings
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.0.0  # PyTorch（用于 HuggingFace 模型）

# ----------------------------------------------------------------------------
# 社区集成和工具
# ----------------------------------------------------------------------------

# 社区贡献的集成
langchain-community>=0.3.0

# 文本分割器和文档加载器
langchain-text-splitters>=0.3.0

# ----------------------------------------------------------------------------
# 向量数据库（用于 RAG）
# ----------------------------------------------------------------------------

# Chroma - 轻量级向量数据库（  荐用于学习）
chromadb>=0.5.0

# FAISS - Facebook AI 相似度搜索
faiss-cpu>=1.8.0
# 如果有 GPU，使用：faiss-gpu>=1.8.0

# Pinecone - 云向量数据库（需要注册）
langchain-pinecone>=0.2.0
pinecone-client>=6.0.0

# Weaviate（可选）
# weaviate-client>=4.0.0

# ----------------------------------------------------------------------------
# 检索增强（用于 RAG Advanced - Module 14）
# ----------------------------------------------------------------------------

# BM25 检索算法（用于混合搜索）
rank_bm25>=0.2.2

# ----------------------------------------------------------------------------
# 文档处理
# ----------------------------------------------------------------------------

# PDF 处理
pypdf>=5.0.0
# 或使用 PyMuPDF（更快）
# pymupdf>=1.24.0

# Word 文档处理
python-docx>=1.1.0

# Excel 处理
openpyxl>=3.1.0

# Markdown 处理
markdown>=3.6

# HTML 处理
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ----------------------------------------------------------------------------
# 数据验证和模型
# ----------------------------------------------------------------------------

# Pydantic - 数据验证（LangChain 依赖）
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0

# ----------------------------------------------------------------------------
# 环境变量管理
# ----------------------------------------------------------------------------

# 环境变量加载
python-dotenv>=1.0.0

# ----------------------------------------------------------------------------
# 网络请求和 API
# ----------------------------------------------------------------------------

# HTTP 客户端
httpx>=0.27.0
requests>=2.32.0

# 异步支持
aiohttp>=3.10.0

# ----------------------------------------------------------------------------
# 数据库（用于 Checkpointing）
# ----------------------------------------------------------------------------

# SQLite Checkpointer - LangGraph 持久化（必需）
langgraph-checkpoint-sqlite>=3.0.0

# PostgreSQL（可选）
# psycopg2-binary>=2.9.0
# 或使用异步版本
# asyncpg>=0.29.0

# Redis（可选，用于缓存）
# redis>=5.0.0

# ----------------------------------------------------------------------------
# 开发和测试工具
# ----------------------------------------------------------------------------

# Jupyter 笔记本
jupyter>=1.1.0
ipykernel>=6.29.0
notebook>=7.2.0

# 测试框架
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# 代码格式化
black>=24.0.0
isort>=5.13.0

# 代码检查
flake8>=7.0.0
pylint>=3.0.0

# 类型检查
mypy>=1.10.0

# ----------------------------------------------------------------------------
# 实用工具
# ----------------------------------------------------------------------------

# 进度条
tqdm>=4.66.0

# 日期时间处理
python-dateutil>=2.9.0

# JSON 处理增强
orjson>=3.10.0

# YAML 处理
pyyaml>=6.0.0

# 正则表达式增强
regex>=2024.0.0

# ----------------------------------------------------------------------------
# 可视化（可选）
# ----------------------------------------------------------------------------

# 图表绘制
# matplotlib>=3.9.0
# seaborn>=0.13.0

# LangGraph 可视化
# graphviz>=0.20.0

# ----------------------------------------------------------------------------
# 搜索和检索工具（可选）
# ----------------------------------------------------------------------------

# Tavily 搜索 API
# tavily-python>=0.5.0

# Google 搜索
# google-search-results>=2.4.0

# DuckDuckGo 搜索
duckduckgo-search>=6.0.0
ddgs

# Wikipedia
# wikipedia>=1.4.0

# arXiv 论文搜索
# arxiv>=2.1.0

# ----------------------------------------------------------------------------
# 性能优化（可选）
# ----------------------------------------------------------------------------

# 更快的 JSON 解析
# ujson>=5.10.0

# 更快的 YAML 解析
# ruamel.yaml>=0.18.0

# ----------------------------------------------------------------------------
# 注意事项
# ----------------------------------------------------------------------------
# 1. 版本号使用语义化版本控制
# 2. 核心包固定主版本号，避免破坏性更新
# 3. 可选包已注释，按需取消注释
# 4. 定期更新依赖以获取安全补丁和新功能
# 5. 在生产环境中建议锁定具体版本号
# ============================================================================

# 如果需要锁定所有依赖的具体版本，运行：
# pip freeze > requirements.lock
